refactor: config 플랫폼 별 패키지로 이동 #3
