+++ src/main/java/flab/project/jobfinder/config/jobkorea/JobKoreaPropertiesConfig.java
@@ -1,5 +1,6 @@
-package flab.project.jobfinder.config;
+package flab.project.jobfinder.config.jobkorea;
 
+import flab.project.jobfinder.config.PropertiesConfig;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
