+++ src/main/java/flab/project/jobfinder/config/rocketpunch/RocketPunchPropertiesConfig.java
@@ -1,5 +1,6 @@
-package flab.project.jobfinder.config;
+package flab.project.jobfinder.config.rocketpunch;
 
+import flab.project.jobfinder.config.PropertiesConfig;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
