+++ src/main/java/flab/project/jobfinder/service/RocketPunchJobFindService.java
@@ -1,6 +1,6 @@
 package flab.project.jobfinder.service;
 
-import flab.project.jobfinder.config.RocketPunchPropertiesConfig;
+import flab.project.jobfinder.config.rocketpunch.RocketPunchPropertiesConfig;
 import flab.project.jobfinder.dto.DetailedSearchDto;
 import flab.project.jobfinder.dto.RecruitDto;
 import flab.project.jobfinder.dto.RecruitPageDto;
