+++ src/test/java/flab/project/jobfinder/service/crawler/RocketPunchCrawlerServiceTest.java
@@ -1,6 +1,6 @@
 package flab.project.jobfinder.service.crawler;
 
-import flab.project.jobfinder.config.RocketPunchPropertiesConfig;
+import flab.project.jobfinder.config.rocketpunch.RocketPunchPropertiesConfig;
 import flab.project.jobfinder.dto.DetailedSearchDto;
 import flab.project.jobfinder.service.crawler.generator.RocketPunchQueryParamGenerator;
 import lombok.extern.slf4j.Slf4j;
